requirements.txt
djangorestframework
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-bootstrap4==1.1.1
idna==2.8
kombu==4.6.8
oauthlib==3.1.0
pika==1.1.0
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
